✏️ useAppContext Hook Separation
