+++ web/aproveme-web/src/hooks/use-app-context.ts
@@ -1,0 +1,11 @@
+import { useContextSelector } from 'use-context-selector'
+
+import { AppContext } from '@/contexts/app-context'
+
+export const useAppContext = () => {
+  const context = useContextSelector(AppContext, (context) => context)
+  if (context === undefined) {
+    throw new Error('useAppContext must be used within an AppProvider')
+  }
+  return context
+}
